second round of changes to accommodate TypedRow - change from Integer to Long, switch some typed strings to rawStrings

+++ services_app/src/androidTest/java/org/opendatakit/database/service/OdkDatabaseBoolTest.java
@@ -91,14 +91,14 @@ public class OdkDatabaseBoolTest extends OdkDatabaseTestAbstractBase {
 
    private void verifyRowTestSet1(TypedRow row) {
       assertEquals(row.getDataByKey(COL_STRING_ID), TEST_STR_1);
-      assertEquals(row.getDataByKey(COL_INTEGER_ID), Integer.valueOf(TEST_INT_1));
+      assertEquals(row.getDataByKey(COL_INTEGER_ID), Long.valueOf(TEST_INT_1));
       assertEquals(row.getDataByKey(COL_BOOL_ID), Boolean.valueOf(TEST_BOOL_1));
    }
 
    private void verifyRowTestSet2(TypedRow row) {
       assertEquals(row.getDataByKey(COL_STRING_ID), TEST_STR_2);
       assertEquals(row.getDataByKey(COL_INTEGER_ID),
-          Integer.valueOf(TEST_INT_2));
+          Long.valueOf(TEST_INT_2));
       assertEquals(row.getDataByKey(COL_BOOL_ID),
           Boolean.valueOf(TEST_BOOL_2));
    }
@@ -106,7 +106,7 @@ public class OdkDatabaseBoolTest extends OdkDatabaseTestAbstractBase {
    private void verifyRowTestSeti(TypedRow row, int i) {
       assertEquals(row.getDataByKey(COL_STRING_ID), TEST_STR_i + i);
       assertEquals(row.getDataByKey(COL_INTEGER_ID),
-          Integer.valueOf(TEST_INT_i + i));
+          Long.valueOf(TEST_INT_i + i));
       assertEquals(row.getDataByKey(COL_BOOL_ID),
           Boolean.valueOf((i%2 == 0)));
    }
@@ -234,7 +234,7 @@ public class OdkDatabaseBoolTest extends OdkDatabaseTestAbstractBase {
 
             TypedRow row = new TypedRow(results.getRowAtIndex(0), columns);
             assertEquals(row.getDataByKey(COL_STRING_ID), TEST_STR_1);
-            assertEquals(row.getDataByKey(COL_INTEGER_ID), Integer.valueOf(changeValue));
+            assertEquals(row.getDataByKey(COL_INTEGER_ID), Long.valueOf(changeValue));
             assertEquals(row.getDataByKey(COL_BOOL_ID), Boolean.valueOf(TEST_BOOL_1));
 
          } catch (ServicesAvailabilityException e) {
